Add Tools module
Move Renderer and which helper method to Tools module

Restore Graph rendering extension
