+++ Sources/Tools/which.swift
@@ -1,0 +1,27 @@
+import Foundation
+
+func which(_ command: String) throws -> URL {
+    let task = Process()
+    let url = URL(fileURLWithPath: "/usr/bin/which")
+    if #available(OSX 10.13, *) {
+        task.executableURL = url
+    } else {
+        task.launchPath = url.path
+    }
+
+    task.arguments = [command.trimmingCharacters(in: .whitespacesAndNewlines)]
+
+    let pipe = Pipe()
+    task.standardOutput = pipe
+    if #available(OSX 10.13, *) {
+        try task.run()
+    } else {
+        task.launch()
+    }
+
+    task.waitUntilExit()
+
+    let data = pipe.fileHandleForReading.readDataToEndOfFile()
+    let string = String(data: data, encoding: .utf8)!.trimmingCharacters(in: .whitespacesAndNewlines)
+    return URL(fileURLWithPath: string)
+}
